Improve exception messages when reverse() lookup fails.
In some cases we can tell what the likely intended result of the lookup
was, and put the semi-matching patterns into the exception message to
make it easier for the developer to work out what they did wrong. In
other cases, particularly in tests, the complete reverse map is not
included in the output, so debugging is harder than necessary.