Split django.db.backends sql_create_model.
Large monolithic functions are very hard to override parts of without
copying and pasting large chunks of code, which is fragile and not DRY.

Split this method into multiple, smaller methods which are easier for
database backends to override and extend. We need this to support creating
UNLOGGED tables in the Postgres backend.